@@ -1,8 +1,8 @@
 package com.workshop.universityannouncementsboard
 
-import android.support.v4.widget.SwipeRefreshLayout
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import com.workshop.universityannouncementsboard.domain.MainPresenter
 import com.workshop.universityannouncementsboard.model.Announcement
 import com.workshop.universityannouncementsboard.presentation.*
