@@ -1,6 +1,6 @@
 package com.workshop.universityannouncementsboard.util
 
-import android.support.v7.widget.RecyclerView
 import android.view.View
+import androidx.recyclerview.widget.RecyclerView
 
 class BaseViewHolder(val view: View) : RecyclerView.ViewHolder(view)
