@@ -1,7 +1,7 @@
 package com.workshop.universityannouncementsboard.util
 
-import android.support.annotation.LayoutRes
 import android.view.*
+import androidx.annotation.LayoutRes
 import kotlinx.android.extensions.*
 
 abstract class ItemAdapter(@LayoutRes open val layoutId: Int) : LayoutContainer {
